Rearranged models a bit and pulled in some models from the exceptionless client
